240202 upd

+++ src/main/java/com/example/demo/service/CommentServiceImpl.java
@@ -2,6 +2,7 @@ package com.example.demo.service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -47,9 +48,18 @@ public class CommentServiceImpl implements CommentService {
 	}
 
 	@Override
-	public void remoeve(int no) { // 댓글 단건 삭제이므로 그냥 no만 받아오면 됨.
-
+	public boolean remove(int no) { // 댓글 단건 삭제이므로 그냥 no만 받아오면 됨.
+		
+		//해당 댓글이 있는지 확인
+		Optional<Comment> comment = repository.findById(no);
+		
+		//없다면  false 반환
+		if (comment.isEmpty()) {
+			return false;
+		}
+		//있다면 댓글 삭제 후 true 반환
 		repository.deleteById(no);
+		return true;
 	}
 
 }
